+++ ProjectProb/app/src/main/java/com/sharktech/projectprob/models/VariableObject.java
@@ -83,8 +83,8 @@ public class VariableObject implements IVariable {
         }
 
         @Override
-        public Float asNumber() {
-            return 1f;
+        public Double asNumber() {
+            return 1d;
         }
 
         @Override
